tonic: refactor example
